MdeModulePkg: CodeQL Fixes.
Includes changes across the repo for the following CodeQL rules:
- cpp/comparison-with-wider-type
- cpp/overflow-buffer
- cpp/redundant-null-check-param
- cpp/uselesstest

+++ MdeModulePkg/Bus/Sd/EmmcBlockIoPei/EmmcHci.c
@@ -934,7 +934,7 @@ BuildAdmaDescTable (
   EFI_PHYSICAL_ADDRESS  Data;
   UINT64                DataLen;
   UINT64                Entries;
-  UINT32                Index;
+  UINT64                Index;
   UINT64                Remaining;
   UINT32                Address;
